poke-backend/pyproject.toml
[project]
name = "weave-backend"
version = "0.1.0"
description = "Weave onboarding backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "langgraph",
    "langchain",
    "langchain-anthropic",
    "composio-langchain==0.8.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]

[project.scripts]
weave-server = "main:main"
